+++ mediation/Tapjoy/source/plugin/Assets/GoogleMobileAds/Mediation/Tapjoy/Common/ITapjoyClient.cs
@@ -1,4 +1,4 @@
-﻿// Copyright 2018 Google LLC
+// Copyright 2018 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
 
 using System;
 
-namespace GoogleMobileAds.Common.Mediation.Tapjoy
+namespace GoogleMobileAds.Mediation.Tapjoy.Common
 {
     public interface ITapjoyClient
     {