+++ mediation/Tapjoy/source/plugin/Assets/GoogleMobileAds/Mediation/Tapjoy/Platforms/Android/TapjoyClient.cs
@@ -1,4 +1,4 @@
-﻿// Copyright 2018 Google LLC
+// Copyright 2018 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -17,9 +17,9 @@
 using UnityEngine;
 using System.Reflection;
 
-using GoogleMobileAds.Common.Mediation.Tapjoy;
+using GoogleMobileAds.Mediation.Tapjoy.Common;
 
-namespace GoogleMobileAds.Android.Mediation.Tapjoy
+namespace GoogleMobileAds.Mediation.Tapjoy.Android
 {
     public class TapjoyClient : ITapjoyClient
     {