+++ mediation/Tapjoy/source/plugin/Assets/GoogleMobileAds/Mediation/Tapjoy/Platforms/iOS/Externs.cs
@@ -1,4 +1,4 @@
-﻿// Copyright 2018 Google LLC
+// Copyright 2018 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -17,7 +17,7 @@
 using System;
 using System.Runtime.InteropServices;
 
-namespace GoogleMobileAds.iOS.Mediation.Tapjoy
+namespace GoogleMobileAds.Mediation.Tapjoy.iOS
 {
     // Externs used by the iOS component.
     internal class Externs