+++ mediation/Tapjoy/source/plugin/Assets/GoogleMobileAds/Mediation/Tapjoy/Platforms/iOS/TapjoyClient.cs
@@ -1,4 +1,4 @@
-﻿// Copyright 2018 Google LLC
+// Copyright 2018 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -16,9 +16,9 @@
 
 using UnityEngine;
 
-using GoogleMobileAds.Common.Mediation.Tapjoy;
+using GoogleMobileAds.Mediation.Tapjoy.Common;
 
-namespace GoogleMobileAds.iOS.Mediation.Tapjoy
+namespace GoogleMobileAds.Mediation.Tapjoy.iOS
 {
     public class TapjoyClient : ITapjoyClient
     {